refactor(google-meet): split Meet API client
